Refactor grouping of syntax highlighting items
Syntax highlighting items (e.g. erlang-font-lock-keywords-vars for
highlighting of variables) are grouped into levels, in order to allow
the user to choose how much highlighting s/he wants. The items need to
appear in a certain order to work properly. In the old implementation,
it was only possible to order items within a group. This led to
impossible to fix bugs (e.g. when calling an external function whose
name clashed with a builtin; 'modul:get(a)'). The new implementation
is backwards compatible, except for fixing the bug mentioned
previously.