Fix highlighting of single-quioted atoms
Single quoted atoms are highlighted by syntactic analysis. In the old
implementation, this led to such atoms being highlighted as strings.
This patch introduces a face for atoms (defaulting to 'default') that is
used for single-quoted strings, i.e. atoms.
A side-effect of this patch is that atoms can be highlighted in a face of
the users choice.